+++ Whono/src/Client/Game/Objects/StaticImageObject.java
@@ -6,7 +6,7 @@ import java.awt.*;
 
 public class StaticImageObject extends GameObject
 {
-	public StaticImageObject(long ID, int x, int y, int width, int height, Image image)
+	public StaticImageObject(long ID, float x, float y, int width, int height, Image image)
 	{
 		super(ID, x, y, width, height, image);
 	}
